requirements.txt
python>=3.9
openai>=1.0.0
together>=0.1.0
transformers>=4.30.0
datasets>=2.10.0
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
torch>=2.0.0
pyyaml>=6.0
python-dotenv>=1.0.0
tqdm>=4.65.0
scipy>=1.10.0
joblib>=1.3.0
